# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API documentation for my application",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate user and return token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jill@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "passwordabcd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "16|ktvP21jNfdjZsrcwSJKTYe93TXBNnco74o727nWn99edd06f"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registers a new user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2021-01-01T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2021-01-01T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "16|ktvP21jNfdjZsrcwSJKTYe93TXBNnco74o727nWn99edd06f"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out",
                "description": "Logs out the authenticated user",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/cars": {
            "get": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Get a list of cars",
                "operationId": "getAllCars",
                "responses": {
                    "200": {
                        "description": "A list of cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Create a car",
                "operationId": "ucreateCar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manufacturing"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string",
                                        "example": "Hundai"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Tucson"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": "2024"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "45000"
                                    },
                                    "manufacturing": {
                                        "type": "integer",
                                        "example": "2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Get a single car",
                "operationId": "getCarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Update a car",
                "operationId": "updateCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand",
                                    "model",
                                    "year",
                                    "price",
                                    "manufacture"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string",
                                        "example": "Hundai"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Tucson"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": "2024"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "45000"
                                    },
                                    "manufacture": {
                                        "type": "integer",
                                        "example": "2020"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Record not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car_CRUD"
                ],
                "summary": "Delete a car",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/properties": {
            "get": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Get a list of properties",
                "operationId": "getAllProperties",
                "responses": {
                    "200": {
                        "description": "A list of properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Property"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Property_CRUD"
                ],
                "summary": "Create a property",
                "operationId": "storeProperty",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "manufacturing"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "House"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Florida"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "455000"
                                    },
                                    "manufacturing": {
                                        "type": "integer",
                                        "example": "2018"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "User_CRUD"
                ],
                "summary": "Get a list of users including their cars and properties",
                "operationId": "getAllUsers",
                "responses": {
                    "200": {
                        "description": "A list of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User_CRUD"
                ],
                "summary": "Create a user",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Peter"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "asdasdasd"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "title": "Car",
                "required": [
                    "id",
                    "brand",
                    "model",
                    "year",
                    "price",
                    "manufacturing"
                ],
                "properties": {
                    "id": {
                        "description": "Car ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "brand": {
                        "description": "Car's brand",
                        "type": "string"
                    },
                    "model": {
                        "description": "Car's model",
                        "type": "string"
                    },
                    "year": {
                        "description": "Car's year",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Car's price",
                        "type": "integer"
                    },
                    "manufacturing": {
                        "description": "Car's manufacturing",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Property": {
                "title": "Property",
                "required": [
                    "id",
                    "type",
                    "address",
                    "price",
                    "manufacturing"
                ],
                "properties": {
                    "id": {
                        "description": "Property ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "description": "Property's type",
                        "type": "string"
                    },
                    "address": {
                        "description": "Property's address",
                        "type": "string"
                    },
                    "price": {
                        "description": "Property's price",
                        "type": "integer"
                    },
                    "manufacturing": {
                        "description": "Property's manufacturing",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "User's password",
                        "type": "string",
                        "format": "password"
                    },
                    "role": {
                        "description": "User's role",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}